backend_flask/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.8.2
botocore==1.35.56
certifi==2024.8.30
click==8.1.7
Deprecated==1.2.14
distro==1.9.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Limiter==3.8.0
Flask-SQLAlchemy==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.0
jmespath==1.0.1
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.1
psycopg2==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
rich==13.9.4
s3transfer==0.10.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.2
wrapt==1.16.0
